Make checksums file usable by common tools
Adding flag --bsd makes rhash output checksums in a format compatible
with common tools, like sha256sum, and makes it obvious what algorithm a
particular checksum was calculated for, at the expense of making the
checksums file larger. This allows for the checksums file to be used
directly by tools understanding traditional format, like Ansible or
Salt, and  obviates the existence of extract-checksum.sh script.