Allow Shebang (`#!`) line #1851
Fairly dirty way of allowing shebang lines

This snippet will read the first two characters of a file passed in to
`yq` and remove the line if it begins with "`#!`". This will allow `yq`
to continue parsing it as though the shebang line was not there.

I was not able to find canonical documentation that the shebang is
limited to a single line or cannot include continuation characters.
However these seem to be accepted truths. See:

https://linux.die.net/man/2/execve

https://en.wikipedia.org/wiki/Shebang_(Unix)#Syntax